Patched to have submodule caching optimizations.
This change allows the submodules to be cached in a repository so that
traversal through all of them is not O(N^2).  Also updated
`git_submodule_status` to stop refreshing expensive information while in this
caching mode.

I have prepared a patch to submit upstream for libgit2; this change is just for
use internally while we wait for (a) me to get permission to contribute to
libgit2, (b) them to take the patch and (c) nodegit to update to that patched
version.